+++ pkg/commands/domainv1/tools/doc.go
@@ -1,0 +1,2 @@
+// Package tools contains commands for interacting with tools for domains.
+package tools